+++ hat/assets/js/apps/Iaso/domains/entities/components/BeneficiaryBaseInfo.tsx
@@ -9,8 +9,8 @@ import {
 import React, { FunctionComponent, useMemo } from 'react';
 // @ts-ignore
 import { useSafeIntl } from 'bluesquare-components';
-import MESSAGES from '../../messages';
-import { useGetPossibleFields } from '../../../forms/hooks/useGetPossibleFields';
+import MESSAGES from '../messages';
+import { useGetPossibleFields } from '../../forms/hooks/useGetPossibleFields';
 import { useGetFields } from '../hooks/useGetFields';
 import { useGetBeneficiaryTypesDropdown } from '../hooks/requests';
