+++ hat/assets/js/apps/Iaso/domains/entities/index.tsx
@@ -11,10 +11,10 @@ import {
     useSafeIntl,
 } from 'bluesquare-components';
 
-import { TableWithDeepLink } from '../../../components/tables/TableWithDeepLink';
-import TopBar from '../../../components/nav/TopBarComponent';
+import { TableWithDeepLink } from '../../components/tables/TableWithDeepLink';
+import TopBar from '../../components/nav/TopBarComponent';
 import { Filters } from './components/Filters';
-import DownloadButtonsComponent from '../../../components/DownloadButtonsComponent';
+import DownloadButtonsComponent from '../../components/DownloadButtonsComponent';
 import {
     useGetBeneficiariesPaginated,
     useGetBeneficiariesApiParams,
@@ -22,12 +22,12 @@ import {
 } from './hooks/requests';
 
 import { useColumns, baseUrl, defaultSorted } from './config';
-import MESSAGES from '../messages';
+import MESSAGES from './messages';
 
-import { redirectTo } from '../../../routing/actions';
+import { redirectTo } from '../../routing/actions';
 import { ListMap } from './components/ListMap';
 
-import { MENU_HEIGHT_WITH_TABS } from '../../../constants/uiConstants';
+import { MENU_HEIGHT_WITH_TABS } from '../../constants/uiConstants';
 
 const useStyles = makeStyles(theme => ({
     ...commonStyles(theme),
