--- hat/assets/js/apps/Iaso/domains/entities/types/entity.ts
@@ -1,10 +1,0 @@
-export type Entity = {
-    id: number;
-    uuid: string;
-    name: string;
-    created_at: number;
-    updated_at: number;
-    attributes: number;
-    entity_type: number;
-    entity_type_name: string;
-};
