+++ hat/assets/js/apps/Iaso/domains/entities/types/fields.ts
@@ -1,4 +1,4 @@
-import { FieldType } from '../../../forms/types/forms';
+import { FieldType } from '../../forms/types/forms';
 
 export type Field = {
     label: string;
