--- hat/assets/js/apps/Iaso/domains/entities/types/paginatedEntities.ts
@@ -1,6 +1,0 @@
-import { Entity } from './entity';
-import { Pagination } from '../../../types/table';
-
-export interface PaginatedEntities extends Pagination {
-    entities: Array<Entity>;
-}
